third_party/WebKit/LayoutTests/fast/canvas/canvas-after-destroy-iframe-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
check if re-parenting canvas keeps its context.
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
check if detaching canvas loses its context.
PASS data[0] is 255
PASS data[1] is 0
PASS data[2] is 0
PASS data[0] is 0
PASS data[1] is 0
PASS data[2] is 0
